Move the token hash method to incoming token
